Move TraceLocation and TraceMarker to tracing
