+++ core/index.ts
@@ -5,6 +5,7 @@ import { TestFixture } from "./test-fixture";
 import { TestLoader } from "./test-loader";
 import { TestOutputStream } from "./test-output-stream";
 import { TestSet } from "./test-set";
+import { TraceLocation, TraceMarker } from "./tracing";
 
 export {
     Expect,
@@ -14,5 +15,7 @@ export {
     TestFixture,
     TestLoader,
     TestOutputStream,
-    TestSet
+    TestSet,
+    TraceLocation,
+    TraceMarker
 };
